feat: add active class to chat row
This commit adds an active class to the chat row when the user is on the chat page.

This is done by adding a useEffect hook that checks if the pathname includes the chat id.

This is done to make it easier for the user to see which chat they are currently on.